Add isOnFirstLine and isOnLastLine to selection module.
