+++ src/device/getElementPaddings.ts
@@ -1,3 +1,5 @@
 /** Gets the padding of an element as an array of numbers. */
-export const getElementPaddings = (element: HTMLElement): number[] =>
+const getElementPaddings = (element: HTMLElement): number[] =>
   window.getComputedStyle(element, null).getPropertyValue('padding').split('px ').map(Number)
+
+export default getElementPaddings
