+++ src/shortcuts/cursorDown.tsx
@@ -1,22 +1,14 @@
 import React from 'react'
 import { Key } from 'ts-key-enum'
-import { parentOf, getElementPaddings, headValue, pathToContext } from '../util'
+import { parentOf, headValue, pathToContext } from '../util'
 import { attributeEquals } from '../selectors'
 import { cursorDown } from '../action-creators'
 import scrollCursorIntoView from '../device/scrollCursorIntoView'
+import * as selection from '../device/selection'
 import { Dispatch, Icon as IconType, Shortcut } from '../@types'
 
 // import directly since util/index is not loaded yet when shortcut is initialized
 import { throttleByAnimationFrame } from '../util/throttleByAnimationFrame'
-
-interface SelectionAttributes {
-  rangeY: number
-  rangeHeight: number
-  baseNodeY: number
-  baseNodeHeight: number
-  paddingTop: number
-  paddingBottom: number
-}
 
 // eslint-disable-next-line jsdoc/require-jsdoc
 const Icon = ({ fill = 'black', size = 20, style }: IconType) => (
@@ -37,44 +29,6 @@ const Icon = ({ fill = 'black', size = 20, style }: IconType) => (
   </svg>
 )
 
-/**
- * Returns selection configuration if exists, null otherwise.
- */
-const getSelectionAttributes = (): SelectionAttributes | null => {
-  const selection = window.getSelection()
-  if (!selection) return null
-
-  const { anchorNode: baseNode, rangeCount } = selection
-  if (rangeCount === 0) return null
-
-  const clientRects = selection.getRangeAt(0).getClientRects()
-  if (!clientRects?.length) return null
-
-  const { y: rangeY, height: rangeHeight } = clientRects[0]
-  if (!rangeY) return null
-
-  const baseNodeParentEl = baseNode?.parentElement as HTMLElement
-  if (!baseNodeParentEl) return null
-
-  const { y: baseNodeY, height: baseNodeHeight } = baseNodeParentEl.getClientRects()[0]
-  const [paddingTop, , paddingBottom] = getElementPaddings(baseNodeParentEl)
-
-  const isMultiline = Math.abs(rangeY - baseNodeY - paddingTop) > 0
-  return isMultiline ? { rangeY, rangeHeight, baseNodeY, baseNodeHeight, paddingTop, paddingBottom } : null
-}
-
-/** Returns true if the selection is on the last line of an editable. */
-const isSelectionOnLastLine = ({
-  rangeY,
-  rangeHeight,
-  baseNodeY,
-  baseNodeHeight,
-  paddingTop,
-  paddingBottom,
-}: SelectionAttributes) => {
-  return rangeY + rangeHeight > baseNodeY + baseNodeHeight - paddingTop - paddingBottom - 5
-}
-
 const cursorDownShortcut: Shortcut = {
   id: 'cursorDown',
   label: 'Cursor Down',
@@ -93,9 +47,8 @@ const cursorDownShortcut: Shortcut = {
     const isProseMode = isProseView && (window.getSelection()?.focusOffset as number) < headValue(cursor).length - 1
     if (isProseMode) return false
 
-    const selectionAttributes = getSelectionAttributes()
-    // use default browser if there is a valid selection and it's not on the last line of a multi-line editable
-    return !selectionAttributes || isSelectionOnLastLine(selectionAttributes)
+    // use default browser behavior (i.e. caret down) if there is a valid selection and it's not on the last line of a multi-line editable
+    return selection.isOnLastLine()
   },
   exec: throttleByAnimationFrame((dispatch: Dispatch) => {
     dispatch(cursorDown())
