+++ src/util/index.ts
@@ -29,7 +29,6 @@ export { formatNumber } from './formatNumber'
 export { getAlgoliaApiKey } from './getAlgoliaApiKey'
 export { getContextMap } from './getContextMap'
 export { getDepth } from './getDepth'
-export { getElementPaddings } from './getElementPaddings'
 export { getPublishUrl } from './getPublishUrl'
 export { getQueryParam } from './getQueryParam'
 export { getUserRef } from './getUserRef'
